requirements.txt
blinker==1.8.1
click==8.1.7
distlib==0.3.8
filelock==3.14.0
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.21.1
packaging==24.0
platformdirs==4.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
upgrade-requirements==1.7.0
virtualenv==20.26.1
Werkzeug==3.0.2